backend/requirements.txt
fastapi
uvicorn
langgraph
langchain-openai
langchain-community
langchain-experimental
tavily-python
langserve[all]
pydantic
python-dotenv
yfinance
pandas
numpy
matplotlib